+++ src/main/g8/app/controllers/actions/DataRetrievalAction.scala
@@ -13,11 +13,8 @@ class DataRetrievalActionImpl @Inject()(
 
   override protected def transform[A](request: IdentifierRequest[A]): Future[OptionalDataRequest[A]] = {
 
-    sessionRepository.get(request.identifier).map {
-      case None =>
-        OptionalDataRequest(request.request, request.identifier, None)
-      case Some(userAnswers) =>
-        OptionalDataRequest(request.request, request.identifier, Some(userAnswers))
+    sessionRepository.get(request.userId).map {
+      OptionalDataRequest(request.request, request.userId, _)
     }
   }
 }