+++ src/main/g8/app/viewmodels/RepeaterAnswerRow.scala
@@ -1,3 +1,0 @@
-package viewmodels
-
-case class RepeaterAnswerRow (answer: String, changeUrl: String, deleteUrl: String)