+++ src/main/g8/app/viewmodels/Section.scala
@@ -1,3 +1,0 @@
-package viewmodels
-
-trait Section