+++ src/main/g8/app/viewmodels/WithCssClass.scala
@@ -1,0 +1,5 @@
+package viewmodels
+
+abstract class WithCssClass(className: String) {
+  override val toString: String = className
+}