+++ src/main/g8/app/viewmodels/WithName.scala
@@ -1,0 +1,5 @@
+package viewmodels
+
+abstract class WithName(name: String) {
+  override val toString: String = name
+}