+++ src/main/g8/app/viewmodels/govuk/package.scala
@@ -1,0 +1,19 @@
+package viewmodels
+
+package object govuk {
+
+  object all
+    extends ImplicitConversions
+      with BackLinkFluency
+      with ButtonFluency
+      with CheckboxFluency
+      with DateFluency
+      with ErrorSummaryFluency
+      with FieldsetFluency
+      with HintFluency
+      with InputFluency
+      with LabelFluency
+      with RadiosFluency
+      with SummaryListFluency
+      with TagFluency
+}