+++ src/main/g8/project/AppDependencies.scala
@@ -5,27 +5,25 @@ object AppDependencies {
 
   val compile = Seq(
     play.sbt.PlayImport.ws,
-    "org.reactivemongo" %% "play2-reactivemongo"            % "0.20.11-play27",
-    "org.reactivemongo" %% "reactivemongo-play-json-compat" % "0.20.11-play27",
-    "com.typesafe.play" %% "play-iteratees"                 % "2.6.1",
-    "uk.gov.hmrc"       %% "govuk-template"                 % "5.57.0-play-27",
-    "uk.gov.hmrc"       %% "play-ui"                        % "8.12.0-play-27",
+    "uk.gov.hmrc"       %% "play-frontend-hmrc"             % "0.57.0-play-27",
     "uk.gov.hmrc"       %% "play-conditional-form-mapping"  % "1.5.0-play-26",
-    "uk.gov.hmrc"       %% "bootstrap-frontend-play-27"     % "4.0.0",
-    "uk.gov.hmrc"       %% "play-language"                  % "4.10.0-play-27"
+    "uk.gov.hmrc"       %% "bootstrap-frontend-play-27"     % "4.2.0",
+    "uk.gov.hmrc"       %% "play-language"                  % "4.10.0-play-27",
+    "uk.gov.hmrc.mongo" %% "hmrc-mongo-play-27"             % "0.49.0"
   )
 
   val test = Seq(
-    "org.scalatest"               %% "scalatest"          % "3.2.0",
-    "org.scalatestplus"           %% "scalacheck-1-14"    % "3.1.2.0",
-    "org.scalatestplus"           %% "mockito-3-3"        % "3.1.2.0",
-    "org.scalatestplus.play"      %% "scalatestplus-play" % "4.0.0",
-    "org.pegdown"                 %  "pegdown"            % "1.6.0",
-    "org.jsoup"                   %  "jsoup"              % "1.10.3",
-    "com.typesafe.play"           %% "play-test"          % PlayVersion.current,
-    "org.mockito"                 %% "mockito-scala"      % "1.16.0",
-    "org.scalacheck"              %% "scalacheck"         % "1.14.3",
-    "com.vladsch.flexmark"        %  "flexmark-all"       % "0.35.10" // Required to stay at this version - see https://github.com/scalatest/scalatest/issues/1736
+    "org.scalatest"           %% "scalatest"               % "3.2.0",
+    "org.scalatestplus"       %% "scalacheck-1-14"         % "3.1.2.0",
+    "org.scalatestplus"       %% "mockito-3-3"             % "3.1.2.0",
+    "org.scalatestplus.play"  %% "scalatestplus-play"      % "4.0.0",
+    "org.pegdown"             %  "pegdown"                 % "1.6.0",
+    "org.jsoup"               %  "jsoup"                   % "1.10.3",
+    "com.typesafe.play"       %% "play-test"               % PlayVersion.current,
+    "org.mockito"             %% "mockito-scala"           % "1.16.0",
+    "org.scalacheck"          %% "scalacheck"              % "1.14.3",
+    "uk.gov.hmrc.mongo"       %% "hmrc-mongo-test-play-27" % "0.47.0",
+    "com.vladsch.flexmark"    %  "flexmark-all"            % "0.35.10" // Required to stay at this version - see https://github.com/scalatest/scalatest/issues/1736
   ).map(_ % "test, it")
 
   def apply(): Seq[ModuleID] = compile ++ test