+++ src/main/g8/test/controllers/actions/SessionActionSpec.scala
@@ -32,7 +32,7 @@ class SessionActionSpec extends SpecBase {
           val result = controller.onPageLoad()(FakeRequest())
 
           status(result) mustBe SEE_OTHER
-          redirectLocation(result).value must startWith(controllers.routes.SessionExpiredController.onPageLoad().url)
+          redirectLocation(result).value must startWith(controllers.routes.JourneyRecoveryController.onPageLoad().url)
         }
       }
     }