+++ src/main/scaffolds/yesNoPage/app/viewmodels/checkAnswers/$className$Summary.scala
@@ -1,0 +1,28 @@
+package viewmodels.checkAnswers
+
+import controllers.routes
+import models.{CheckMode, UserAnswers}
+import pages.$className$Page
+import play.api.i18n.Messages
+import uk.gov.hmrc.govukfrontend.views.viewmodels.summarylist.SummaryListRow
+import viewmodels.govuk.summarylist._
+import viewmodels.implicits._
+
+object $className$Summary  {
+
+  def row(answers: UserAnswers)(implicit messages: Messages): Option[SummaryListRow] =
+    answers.get($className$Page).map {
+      answer =>
+
+        val value = if (answer) "site.yes" else "site.no"
+
+        SummaryListRowViewModel(
+          key     = "$className;format="decap"$.checkYourAnswersLabel",
+          value   = ValueViewModel(value),
+          actions = Seq(
+            ActionItemViewModel("site.change", routes.$className$Controller.onPageLoad(CheckMode).url)
+              .withVisuallyHiddenText(messages("$className;format="decap"$.change.hidden"))
+          )
+        )
+    }
+}